% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.igraph.R
\name{as.igraph}
\alias{as.igraph}
\title{as.igraph}
\usage{
as.igraph(x)
}
\arguments{
\item{x}{the output from the function plot.}
}
\value{
the CoDiNA plot as an igraph object.
}
\description{
Converts the CoDiNA.plot into an igraph object.
}
\examples{
suppressWarnings(RNGversion("3.5.0"))
Nodes = LETTERS[1:10]
Net1 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
Net2 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
Net3 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
DiffNet = MakeDiffNet (Data = list(Net1,Net2,Net3), Code = c('Net1', 'Net2', 'Net3') )
Graph = plot(x = DiffNet,
 layout = NULL, smooth.edges = TRUE,
 path = NULL, MakeGroups = FALSE, Cluster = FALSE,
 legend = TRUE, manipulation = FALSE, sort.by.Phi = FALSE)
x = as.igraph(Graph)

plot(x)

}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
