% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categorize.R
\name{Categorize}
\alias{Categorize}
\title{Categorize}
\usage{
Categorize(M, cutoff = 0.33)
}
\arguments{
\item{M}{A data.frame to be categorized.}

\item{cutoff}{By default, the cutoff is 0.33. If the user wants to use another value, it has to be cited on the description of the used methodology that the cutoff was changed.}
}
\description{
Categorize the links into -1, 0 and 1 given a cutoff.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
\keyword{internal}
