% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_full_base.R
\name{CreateFullBase}
\alias{CreateFullBase}
\title{CreateFullBase}
\usage{
CreateFullBase(x, Code)
}
\arguments{
\item{x}{List of data.frames containig Node.1, Node.2 and the correlation value}

\item{Code}{Name of each one of the networks.}
}
\value{
Returns a list contating: The nodes names and it's correlation values in all networks, 0 if this node is absent.
}
\description{
Joins a set of data.frames, order the nodes names by it's smaller value.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
\keyword{internal}
