% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(m)
}
\arguments{
\item{m}{variable to be normalized in the interval [0,1]}
}
\description{
Normalize a given variable.
}
\examples{
Z = runif(10,-10,10)
normalize(Z)
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
