% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderNames.R
\name{OrderNames}
\alias{OrderNames}
\title{OrderNames}
\usage{
OrderNames(M)
}
\arguments{
\item{M}{data.frame to have the names ordered. Node.1, Node.2 and correlation value.}
}
\value{
a data.table whith Node.1 and Node.2, sorted by the smallest value between both.
}
\description{
Sorts each link's Nodes by the smallest value. Removes links that both nodes are the same.
}
\examples{
Nodes = LETTERS[1:10]
Z = data.frame(Node.1 = sample(Nodes) ,
Node.2 = sample(Nodes), cor = runif(10,-1,1))
OrderNames(Z)
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
