% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoDiNA_Categorize.R
\name{PhiCategory}
\alias{PhiCategory}
\title{PhiCategory}
\usage{
PhiCategory(Base, n, CAT_BASE = CAT_BASE)
}
\arguments{
\item{Base}{data.frame to be categorized}

\item{n}{number of networks to be compared}
}
\description{
Categorize the links into Phi and Phi tilda categories.
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
\keyword{internal}
