% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.CoDiNA}
\alias{print.CoDiNA}
\title{print.CoDiNA}
\usage{
\method{print}{CoDiNA}(x, ...)
}
\arguments{
\item{x}{Output from MakeDiffNet}

\item{\dots}{Additional plotting parameters.}
}
\value{
Print on the screen the number of nodes and edges.
}
\description{
Print on the screen the number of nodes and edges. To see the data.frame, call: data.frame().
}
\examples{
Nodes = LETTERS[1:10]
Net1 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
Net2 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
Net3 = data.frame(Node.1 = sample(Nodes) , Node.2 = sample(Nodes), wTO = runif(10,-1,1))
DiffNet = MakeDiffNet (Data = list(Net1,Net2,Net3), Code = c('Net1', 'Net2', 'Net3') )
print(DiffNet)
}
\author{
Deisy Morselli Gysi <deisy at bioinf.uni-leipzig.de>
}
