% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recode.R
\name{Recode}
\alias{Recode}
\title{Title Recode}
\usage{
Recode(Phi, values)
}
\arguments{
\item{Phi}{Phi categories}

\item{values}{Categorical base}
}
\description{
Title Recode
}
\keyword{internal}
